refactor: rename files and remove unused utils
