+++ apps/diet-app/src/app/calendar-day/page.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Calendar } from '../../client/components/calendar/calendar';
 import Footer from '../../client/components/footer/footer';
-import CaloriesBar from '../../client/components/caloriesBar/caloriesBar';
+import CaloriesBar from '../../client/components/calories-bar/calories-bar';
 import Wrapper from '../../client/components/wrapper/wrapper';
 import Nav from '../../client/components/nav/nav';
-import Meal from '../../client/components/meal/meal';
+import Meal from '../../client/components/single-meal/single-meal';
 
 export default function CalendarDay() {
   return (
