+++ apps/diet-app/src/client/components/calories-bar/calories-bar.tsx
